Update Docker to build 3 different containers
This update gives us 3 different containers that can be used depending
on what the user's needs are. I suspect though that most people are
going to settle on CUDA or ALL.

The first version is MIN, which is the smallest and has:
- PyWPA
- PWA2000

The second version is CUDA, which is bigger than MIN but has CUDA
enabled.

The third is ALL which is CUDA enabled, but also includes:
- Jupyter and Jupyterlab
- iPython
- Tensorflow 2.0
- ROOT